# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Gantt App API",
    "description": "API REST para gestión de proyectos y tareas con diagramas Gantt",
    "contact": {
      "name": "Gantt App Team"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base URL"
    }
  ],
  "paths": {
...
    "/projects": {
      "get": {
        "tags": ["Proyectos"],
        "summary": "Listar proyectos",
        "description": "Lista paginada de proyectos del usuario",
        "operationId": "e66a4b9c561c9e3e0b98d95b6437ff91",
        "parameters": [
          {
            "name": "per_page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "status_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de proyectos"
          },
          "401": {
            "description": "No autenticado"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      },
      "post": {
        "tags": ["Proyectos"],
        "summary": "Crear proyecto",
        "description": "Crea un nuevo proyecto (solo Admin o Super Admin)",
        "operationId": "3e5e24c249cabe95d4c10927fda15c42",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["company_id", "name", "color"],
                "properties": {
                  "company_id": {
                    "type": "integer"
                  },
                  "project_status_id": {
                    "type": "integer"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string",
                    "nullable": true
                  },
                  "color": {
                    "type": "string"
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proyecto creado"
          },
          "403": {
            "description": "No autorizado"
          },
          "422": {
            "description": "Error de validación"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      }
    },
    "/projects/{project}": {
      "get": {
        "tags": ["Proyectos"],
        "summary": "Detalle de proyecto",
        "description": "Devuelve el detalle de un proyecto",
        "operationId": "042e8a7d144e8e7d6f2bc9009f82ae3a",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "include_stats",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del proyecto"
          },
          "404": {
            "description": "No encontrado"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      },
      "delete": {
        "tags": ["Proyectos"],
        "summary": "Eliminar proyecto",
        "description": "Eliminación lógica (soft delete vía status)",
        "operationId": "9d379dcf6b99d1860846be1e3b9520a9",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proyecto eliminado"
          },
          "422": {
            "description": "Estado no permite eliminación"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      },
      "patch": {
        "tags": ["Proyectos"],
        "summary": "Actualizar proyecto",
        "description": "Actualización parcial de un proyecto",
        "operationId": "c8489e0b6f426883dbfc90a81c532bb8",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "project_status_id": {
                    "type": "integer"
                  },
                  "color": {
                    "type": "string"
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date"
                  }
                },
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Proyecto actualizado"
          },
          "422": {
            "description": "Transición de estado inválida"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      }
    },
    "/projects/{project}/restore": {
      "post": {
        "tags": ["Proyectos"],
        "summary": "Restaurar proyecto",
        "description": "Restaura un proyecto eliminado",
        "operationId": "7b25c2a0f064ede37cae606b0cbec69c",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proyecto restaurado"
          },
          "422": {
            "description": "El proyecto no está eliminado"
          }
        },
        "security": [
          {
            "sanctum": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "description": "Token de autenticación Laravel Sanctum",
        "bearerFormat": "JWT",
        "scheme": "bearer"
      }
    }
  },
...
}